Update Get-StakRank.ps1
Refactored the function that retrieved file headers from the first input file matching the input file pattern, splitting it into two functions, one that gets the list of files matching the pattern, the other gets the fields from the first file in that list. This prevents having to traverse the file system twice.

Added a function to read optional roles data from file.

Stubbed out a function for doing the ranking.